refactor(Fias): move fields settings up to the Fias component
